% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{getActiveGeneMethod}
\alias{getActiveGeneMethod}
\title{Active Gene Methods}
\usage{
getActiveGeneMethod(dataset)
}
\value{
Returns the available methods data for availability
}
\description{
Active Gene Methods
}
