% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{getActiveGenesInCellType}
\alias{getActiveGenesInCellType}
\title{Active Genes In All Brain Region For Cell Type}
\usage{
getActiveGenesInCellType(dataset, celltype, method)
}
\value{
Returns a matrix with the active/inactive genes in all brain region, for a given cell type and calculus method.
}
\description{
Active Genes In All Brain Region For Cell Type
}
