% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{getBrainregionsDataSet}
\alias{getBrainregionsDataSet}
\title{Brain Regions in Dataset}
\usage{
getBrainregionsDataSet(dataset)
}
\value{
Returns the brain regions considered in the data set identified by the string in dataset
}
\description{
Brain Regions in Dataset
}
