% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{getCellTypesInDataSet}
\alias{getCellTypesInDataSet}
\title{Cell Types in Dataset}
\usage{
getCellTypesInDataSet(dataset)
}
\value{
Returns the cell types considered in the data set identified by the string in dataset
}
\description{
Cell Types in Dataset
}
