% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{getGenesInCellType}
\alias{getGenesInCellType}
\title{Active and Inactive Genes In All Brain Region For Cell Type}
\usage{
getGenesInCellType(dataset, celltype, method, visibility)
}
\value{
Returns a matrix with the active/inactive genes in all brain region, for a given cell type and calculus method. In rawexpression method, visibility jumps from 0 to 10.
}
\description{
Active and Inactive Genes In All Brain Region For Cell Type
}
