% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{getActiveGenesInBrainRegionInCellType}
\alias{getActiveGenesInBrainRegionInCellType}
\title{Active Genes by Brain Region And Cell Type}
\usage{
getActiveGenesInBrainRegionInCellType(dataset, brainregion, celltype, method)
}
\value{
Returns the gene ids of active genes in the brain region, cell type, for a given calculus method.
}
\description{
Active Genes by Brain Region And Cell Type
}
