% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{getActiveGenesInCellType}
\alias{getActiveGenesInCellType}
\title{Active Genes In Cell Type}
\usage{
getActiveGenesInCellType(dataset, brainregion, celltype, method)
}
\value{
Returns the gene ids of genes active in the brain region, cell type, for a given calculus method.
}
\description{
Active Genes In Cell Type
}
